files and upload have finished

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const { validationResult } = require('express-validator/check');
 
+const fileHelper=require('../util/file');
+
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
@@ -149,6 +151,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if(image){
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then(result => {
@@ -182,7 +185,12 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId).then(product=>{
+    if(!product){
+      return next(new Error("Product not Found..."))
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    Product.deleteOne({ _id: prodId, userId: req.user._id })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
@@ -191,4 +199,8 @@ exports.postDeleteProduct = (req, res, next) => {
       const error=new Error(err);
       error.httpStatusCode=500;
       return next(error);
-    });};
+    });
+  }).catch(err=>{
+    next(err);
+  })
+  };
